add curriculum

+++ application/migrations/20170205022932_logs.php
@@ -34,7 +34,7 @@ class Migration_Logs extends CI_Migration
                 $this->down();
                 $fields = array(
                     'id'         => array(
-                        'type'           => 'MEDIUMINT',
+                        'type'           => 'INT',
                         'constraint'     => 8,
                         'unsigned'       => TRUE,
                         'null'           => FALSE,
